--- packages/manager/src/routes/account/index.tsx
@@ -1,14 +1,0 @@
-import { Outlet } from '@tanstack/react-router';
-import React from 'react';
-
-import { ProductInformationBanner } from 'src/components/ProductInformationBanner/ProductInformationBanner';
-import { SuspenseLoader } from 'src/components/SuspenseLoader';
-
-export const AccountRoutes = () => {
-  return (
-    <React.Suspense fallback={<SuspenseLoader />}>
-      <ProductInformationBanner bannerLocation="Account" />
-      <Outlet />
-    </React.Suspense>
-  );
-};
