+++ packages/manager/src/routes/betas/index.ts
@@ -1,29 +1,29 @@
 import { createRoute } from '@tanstack/react-router';
 
 import { rootRoute } from '../root';
-import { BetaRoutes } from './BetasRoute';
+import { BetasRoute } from './BetasRoute';
 
-const betaRoute = createRoute({
-  component: BetaRoutes,
+const betasRoute = createRoute({
+  component: BetasRoute,
   getParentRoute: () => rootRoute,
   path: 'betas',
 });
 
 const betaLandingRoute = createRoute({
-  getParentRoute: () => betaRoute,
+  getParentRoute: () => betasRoute,
   path: '/',
 }).lazy(() =>
   import('src/features/Betas/BetasLanding').then((m) => m.betasLandingLazyRoute)
 );
 
 const betaSignupRoute = createRoute({
-  getParentRoute: () => betaRoute,
+  getParentRoute: () => betasRoute,
   path: 'signup/$betaId',
 }).lazy(() =>
   import('src/features/Betas/BetaSignup').then((m) => m.betaSignupLazyRoute)
 );
 
-export const betaRouteTree = betaRoute.addChildren([
+export const betaRouteTree = betasRoute.addChildren([
   betaLandingRoute,
   betaSignupRoute,
 ]);
