+++ packages/manager/src/routes/cloudPulse/CloudPulseRoute.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import { SuspenseLoader } from 'src/components/SuspenseLoader';
 
-export const CloudPulseRoutes = () => {
+export const CloudPulseRoute = () => {
   return (
     <React.Suspense fallback={<SuspenseLoader />}>
       <DocumentTitleSegment segment="Cloud Pulse" />
