+++ packages/manager/src/routes/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import { NotFound } from 'src/components/NotFound';
 
-import { accountRouteTree } from './account/AccountRoute';
+import { accountRouteTree } from './account';
 import { betaRouteTree } from './betas';
 import { cloudPulseRouteTree } from './cloudPulse';
 import { databasesRouteTree } from './databases';
