+++ packages/manager/src/routes/longview/LongviewRoute.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { ProductInformationBanner } from 'src/components/ProductInformationBanner/ProductInformationBanner';
 import { SuspenseLoader } from 'src/components/SuspenseLoader';
 
-export const LongviewRoutes = () => {
+export const LongviewRoute = () => {
   return (
     <React.Suspense fallback={<SuspenseLoader />}>
       <ProductInformationBanner bannerLocation="Longview" />
