+++ packages/manager/src/routes/profile/ProfileRoute.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import { SuspenseLoader } from 'src/components/SuspenseLoader';
 
-export const ProfileRoutes = () => {
+export const ProfileRoute = () => {
   return (
     <React.Suspense fallback={<SuspenseLoader />}>
       <Outlet />
